Table remove books button

+++ functions/helpers/remove-document.js
@@ -6,13 +6,21 @@ import { Cross2Icon } from "@radix-ui/react-icons";
 const removeDocument = async (book, collection) => {
   try {
     const user = auth.currentUser;
-    await deleteDoc(doc(db, "users", user.uid, collection, book.bookId));
-    showNotification({
-      title: "Removed",
-      message: `${book.title}`,
-      color: "pink",
-      icon: <Cross2Icon />,
-    });
+    if (!Array.isArray(book)) {
+      await deleteDoc(doc(db, "users", user.uid, collection, book.bookId));
+      showNotification({
+        title: "Removed",
+        message: `${book.title}`,
+        color: "pink",
+        icon: <Cross2Icon />,
+      });
+    } else {
+      showNotification({
+        title: "Removed",
+        message: `${book.length} books`,
+        color: "pink",
+      });
+    }
   } catch (error) {
     showNotification({
       title: "Error",
